requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
Pillow
matplotlib